revert refactor of DelegateTree JsonCodec
