+++ namerd/iface/control-http/src/main/scala/io/buoyant/namerd/iface/StreamingNamerClient.scala
@@ -7,7 +7,6 @@ import com.twitter.finagle._
 import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finagle.naming.NameInterpreter
 import com.twitter.util.{Activity, Closable, Future, Var}
-import io.buoyant.admin.DelegationJsonCodec
 import io.buoyant.admin.names.DelegateApiHandler
 import io.buoyant.admin.names.DelegateApiHandler.JsonDelegateTree
 import io.buoyant.config.JsonStreamParser
@@ -63,7 +62,7 @@ class StreamingNamerClient(
       }
     )
 
-  private[this] val mapper = DelegationJsonCodec.mapper
+  private[this] val mapper = DelegateApiHandler.Codec.mapper
   mapper.registerModule(DtabModule.module)
   private[this] val parser = new JsonStreamParser(mapper)
